+++ packages/app/nam-test/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.16](https://github.com/andremichelle/openDAW/compare/@opendaw/nam-test@0.0.15...@opendaw/nam-test@0.0.16) (2026-06-05)
+
+**Note:** Version bump only for package @opendaw/nam-test
+
+## [0.0.15](https://github.com/andremichelle/openDAW/compare/@opendaw/nam-test@0.0.14...@opendaw/nam-test@0.0.15) (2026-05-22)
+
+**Note:** Version bump only for package @opendaw/nam-test
+
 ## [0.0.14](https://github.com/andremichelle/openDAW/compare/@opendaw/nam-test@0.0.13...@opendaw/nam-test@0.0.14) (2026-05-13)
 
 **Note:** Version bump only for package @opendaw/nam-test