+++ packages/lib/midi/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.66](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-midi@0.0.65...@opendaw/lib-midi@0.0.66) (2026-05-22)
+
+**Note:** Version bump only for package @opendaw/lib-midi
+
 ## [0.0.65](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-midi@0.0.64...@opendaw/lib-midi@0.0.65) (2026-05-13)
 
 **Note:** Version bump only for package @opendaw/lib-midi