+++ packages/studio/forge-boxes/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.94](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-forge-boxes@0.0.93...@opendaw/studio-forge-boxes@0.0.94) (2026-06-05)
+
+**Note:** Version bump only for package @opendaw/studio-forge-boxes
+
+## [0.0.93](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-forge-boxes@0.0.92...@opendaw/studio-forge-boxes@0.0.93) (2026-06-05)
+
+**Note:** Version bump only for package @opendaw/studio-forge-boxes
+
+## [0.0.92](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-forge-boxes@0.0.91...@opendaw/studio-forge-boxes@0.0.92) (2026-06-03)
+
+**Note:** Version bump only for package @opendaw/studio-forge-boxes
+
+## [0.0.91](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-forge-boxes@0.0.90...@opendaw/studio-forge-boxes@0.0.91) (2026-05-22)
+
+**Note:** Version bump only for package @opendaw/studio-forge-boxes
+
 ## [0.0.90](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-forge-boxes@0.0.89...@opendaw/studio-forge-boxes@0.0.90) (2026-05-13)
 
 **Note:** Version bump only for package @opendaw/studio-forge-boxes